+++ apps/airbash-frontend/src/app/services/data.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-import { ApiService } from './api.service';
+import { ApiService } from './services/api.service';
 import { Capture } from './types/capture';
 
 @Injectable({
